agora vai

+++ NutriApp/Database/FoodDB.cs
@@ -1,70 +1,45 @@
-using System.Collections.ObjectModel;
-using NutriApp.Models;
-
 namespace NutriApp.Database;
 
-public class FoodDb
-{
-    public FoodDb()
-    {
+public class FoodDb {
+    public FoodDb() {
         App.Database.CreateTableAsync<FoodModel>().Wait();
     }
-    public async Task<IOrderedEnumerable<FoodModel>> ListarAsync()
-    {
-        var foodListResult = new List<FoodModel>();
-        var foodList = await App.Database.Table<FoodModel>().ToListAsync();
-        foreach (var foodModel in foodList)
-        {
-            foodModel.Nome = foodModel.Nome.Replace("-Não se aplica", "");
-            foodListResult.Add(foodModel);
-        }
-        //TODO - ARRUMAR LISTAGEM
-        //return foodListResult.OrderBy(food => food.Nome);
-        return foodListResult.Take(500).OrderBy(food => food.Nome);
+
+    public async Task<IEnumerable<FoodModel>> FoodListAsync() {
+        return await App.Database.Table<FoodModel>().OrderBy(food => food.Nome).ToListAsync();
     }
 
-    public async Task<int> CadastrarListaAsync(FoodModel foodModel)
-    {
+    public async Task<int> CadastrarListaAsync(FoodModel foodModel) {
         return await App.Database.InsertAsync(foodModel);
     }
-        
-    public async Task<int> AtualizarAsync(FoodModel foodModel)
-    {
+
+    public async Task<int> AtualizarAsync(FoodModel foodModel) {
         return await App.Database.UpdateAsync(foodModel);
     }
 
-    public async Task<int> DeleteTotalAsync(int id)
-    {
+    public async Task<int> DeleteTotalAsync(int id) {
         // Delete a note.
         var foodModel = await ConsultarAsync(id);
-        if (foodModel != null)
-        {
-            return await App.Database.DeleteAsync(foodModel);
-        }
+        if (foodModel != null) return await App.Database.DeleteAsync(foodModel);
 
         return 0;
     }
 
-    public async Task<FoodModel> ConsultarAsync(int id)
-    {
+    public async Task<FoodModel> ConsultarAsync(int id) {
         return await App.Database.Table<FoodModel>()
             .Where(i => i.Id == id)
             .FirstOrDefaultAsync();
     }
-    public async Task<bool> UpdateListAsync(ObservableCollection<FoodModel> listFood)
-    {
-        foreach (var outsideFood in listFood)
-        {
+
+    public async Task<bool> UpdateListAsync(ObservableCollection<FoodModel> listFood) {
+        foreach (var outsideFood in listFood) {
             var food = await ConsultarAsync(outsideFood.Id);
             if (food == null)
-            {
                 await CadastrarListaAsync(outsideFood);
-            }
             else
-            {
                 await AtualizarAsync(outsideFood);
-            }
         }
+
         return true;
     }
 }
